FIX: min function

+++ src/kernels/resample_2d.cu
@@ -49,7 +49,7 @@ extern "C" __global__ void resample_x(
     float right_f = (globalID.x + 1) * delta;
 
     int left_i = static_cast<int>(floor(left_f));
-    int right_i = min(in_width, static_cast<size_t>(ceil(right_f)));
+    int right_i = min(static_cast<int>(in_width), static_cast<int>(ceil(right_f)));
 
     float value = 0.f;
     
@@ -92,7 +92,7 @@ extern "C" __global__ void resample_y(
     float right_f = (globalID.y + 1) * delta;
 
     int left_i = static_cast<int>(floor(left_f));
-    int right_i = min(in_height, static_cast<size_t>(ceil(right_f)));
+    int right_i = min(static_cast<int>(in_height), static_cast<int>(ceil(right_f)));
 
     float value = 0.f;
